Updated GUI to be more responsive.
The program doesn't run quickly on older computers and it can appear to be broken or frozen because of that.  I redid the GUI coding to be more compartmentalized and added elements that will inform the user the program is working.  It also updates them again after it is over to let them know that it is finished.